requirements.txt
asgiref==3.2.9
Django==2.2.12
django-adminlte-ui==1.5.1
django-adminlte2==0.4.3
django-adminlte2-templates==1.3.2
django-appconf==1.0.4
django-crispy-forms==1.7.2
django-cruds-adminlte==0.0.15
django-image-cropping==1.4.0
django-redis==4.12.1
django-select2==7.4.2
django-treebeard==4.3
djangoajax==2.3.7
easy-thumbnails==2.7
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.5
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
